update key regression

+++ include/configure.hpp
@@ -87,7 +87,7 @@ private:
     int _keyBatchSize;
     uint64_t _keyCacheSize;
     uint64_t _keyGenLimitPerSessionKey;
-    uint64_t _keyRegressionMaxTimes;
+    uint32_t _keyRegressionMaxTimes;
 
     //muti thread settings
     int _encodeThreadLimit;
@@ -179,7 +179,7 @@ public:
     int getKeyBatchSize();
     uint64_t getKeyCacheSize();
     uint64_t getKeyGenLimitPerSessionkeySize();
-    uint64_t getKeyRegressionMaxTimes();
+    uint32_t getKeyRegressionMaxTimes();
 
     //muti thread settings
     int
